Attempt to make Component more Java-like
