+++ lib/src/main/java/org/js/cycle/android/Component.java
@@ -1,0 +1,5 @@
+package org.js.cycle.android;
+
+public interface Component {
+  Sinks create(Sources sources);
+}
